projects/serializers.py
from .models import Project
from rest_framework import serializers


class ProjectSerializer(serializers.ModelSerializer):
    class Meta:
        model = Project
        fields = ("id", "name", "title", "description", "beauty_shot", "active_flag", )
