requirements.txt
boto3==1.26.118
botocore==1.29.118
Django==4.2.7
django-anymail==9.1
django-cors-headers==3.14.0
django-imagekit==4.1.0
django-storages==1.13.2
django-structlog==5.0.2
django-templated-email==3.0.1
djangorestframework==3.14.0
djoser==2.2.2
Pillow==10.1.0
requests==2.31.0
pytz==2023.3
pilkit~=2.0
structlog~=23.1.0